mnist_deploy/requirements.txt
flask
numpy
pillow
scipy
